requirements.txt
Flask==1.0.2
Flask-Injector==0.6.2
Flask-RESTful==0.3.5
google-api-python-client==1.7.3
injector==0.12.1
